requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
SQLAlchemy==1.3.20
Werkzeug==1.0.1
flasgger~=0.9.5